src/main/java/com/stefandekanski/weakwrap/processor/WeakWrapProcessor.java
package com.stefandekanski.weakwrap.processor;

import com.google.auto.service.AutoService;
import com.stefandekanski.weakwrap.anotation.WeakWrap;

import javax.annotation.processing.*;
import javax.lang.model.SourceVersion;
import javax.lang.model.element.Element;
import javax.lang.model.element.TypeElement;
import javax.lang.model.util.Elements;
import javax.lang.model.util.Types;
import javax.tools.Diagnostic;
import java.io.IOException;
import java.util.Collections;
import java.util.Set;

@AutoService(Processor.class)
public class WeakWrapProcessor extends AbstractProcessor {

    private Types typeUtils;
    private Elements elementUtils;
    private Filer filer;
    private Messager messager;

    @Override
    public synchronized void init(ProcessingEnvironment processingEnv) {
        super.init(processingEnv);
        typeUtils = processingEnv.getTypeUtils();
        elementUtils = processingEnv.getElementUtils();
        filer = processingEnv.getFiler();
        messager = processingEnv.getMessager();
    }

    @Override
    public Set<String> getSupportedAnnotationTypes() {
        return Collections.singleton(WeakWrap.class.getCanonicalName());
    }

    @Override
    public SourceVersion getSupportedSourceVersion() {
        return SourceVersion.latestSupported();
    }

    @Override
    public boolean process(Set<? extends TypeElement> annotations, RoundEnvironment roundEnv) {
        Set<? extends Element> elements = roundEnv.getElementsAnnotatedWith(WeakWrap.class);
        boolean processed = false;
        try {
            for (Element e : elements) {
                TypeElement typeElement = (TypeElement) e;
                WeakWrapWriter weakWrapWriter = new WeakWrapWriter(typeElement);
                weakWrapWriter.writeWeakWrapperTo(filer);
                processed = true;
            }
        } catch (IOException e) {
            messager.printMessage(Diagnostic.Kind.ERROR, e.getMessage());
        }
        return processed;
    }




}
